incorporate review comments + fix issues

+++ docker/atomic-certs-update.sh
@@ -35,14 +35,14 @@ cd $CERTS_DIR
 
 # Old format, move to using double symlinks
 # also upgrade old containers to use new envoy image
-if [[ -f $CERT_FILE ]]; then
+if [[ -f $CERT_FILE && ! -L $CERT_FILE ]]; then
         OLD_CERTS_DIR=..certs_$(date "+%Y_%m_%d_%H_%M_%S.%s")
         mkdir -p $OLD_CERTS_DIR
         mv $CERT_FILE $OLD_CERTS_DIR/
         mv $KEY_FILE $OLD_CERTS_DIR/
-        ln -sf $OLD_CERTS_DIR ..data
-        ln -sf ..data/$CERT_FILE $CERT_FILE
-        ln -sf ..data/$KEY_FILE $KEY_FILE
+        ln -snf $OLD_CERTS_DIR ..data
+        ln -snf ..data/$CERT_FILE $CERT_FILE
+        ln -snf ..data/$KEY_FILE $KEY_FILE
 
 	for envoyName in $(docker ps --format "{{.Names}}" --filter name="^envoy"); do
 	  envoyPath="/home/ubuntu/envoy/${envoyName#envoy}"
@@ -52,7 +52,9 @@ if [[ -f $CERT_FILE ]]; then
 	  fi
 
 	  # patch envoy.yaml to use sds
-	  sed -i "1 s/^/node:\n  id: ${envoyName}\n  cluster: ${envoyName}/" $envoyPath/envoy.yaml
+	  if ! grep -iq "node:" $envoyPath/envoy.yaml; then
+		  sed -i "1 s/^/node:\n  id: ${envoyName}\n  cluster: ${envoyName}/" $envoyPath/envoy.yaml
+	  fi
 
 	  sed -i '/      tls_context:/,/.key"/c\
       transport_socket:\
@@ -68,7 +70,7 @@ if [[ -f $CERT_FILE ]]; then
 	  cat > $envoyPath/sds.yaml <<EOF
 resources:
 - "@type": "type.googleapis.com/envoy.api.v2.auth.Secret"
-  tls_certificates:
+  tls_certificate:
     certificate_chain:
       filename: "/etc/envoy/certs/envoyTlsCerts.crt"
     private_key:
@@ -76,10 +78,13 @@ resources:
 EOF
 
 	  # stop and start docker with new image
+	  runcmd=$(docker run --rm -v /var/run/docker.sock:/var/run/docker.sock assaflavie/runlike $envoyName)
 	  docker stop $envoyName
 	  docker rm $envoyName
-	  runcmd=$(docker run --rm -v /var/run/docker.sock:/var/run/docker.sock assaflavie/runlike $envoyName)
-	  new_runcmd=($(sed 's/envoy-with-curl.*\? /envoy-with-curl@sha256:9bc06553ad6add6bfef1d8a1b04f09721415975e2507da0a2d5b914c066474df /g' <<< $runcmd))
+	  # mount sds.yaml file
+	  runcmd=$(sed "s|--detach|--volume=${envoyPath}/sds.yaml:/etc/envoy/sds.yaml --detach|g" <<< $runcmd)
+	  # use latest envoy-with-curl docker image
+	  new_runcmd=($(sed 's/envoy-with-curl.*? /envoy-with-curl@sha256:9bc06553ad6add6bfef1d8a1b04f09721415975e2507da0a2d5b914c066474df /g' <<< $runcmd))
 	  echo "$envoyName=>$new_runcmd"
 	  "${new_runcmd[@]}"
 	done
@@ -91,8 +96,13 @@ if [[ -f $CERT_FILE.new ]]; then
         mkdir -p $NEW_CERTS_DIR
         mv $CERT_FILE.new $NEW_CERTS_DIR/$CERT_FILE
         mv $KEY_FILE.new $NEW_CERTS_DIR/$KEY_FILE
-        rm -f ..tmp; ln -sf $NEW_CERTS_DIR ..tmp
+        rm -f ..tmp; ln -snf $NEW_CERTS_DIR ..tmp
         mv -Tf ..tmp ..data
+
+	if [[ ! -L $CERT_FILE ]]; then
+	  ln -snf ..data/$CERT_FILE $CERT_FILE
+	  ln -snf ..data/$KEY_FILE $KEY_FILE
+	fi
 
         # Prune old certs
         find -type d -name "..certs_*" -not -path "./$NEW_CERTS_DIR" | xargs rm -rf
